Add mixin and Setup main
