+++ src/main/resources/META-INF/mods.toml
@@ -71,3 +71,5 @@ description='''${mod_description}'''
 # stop your mod loading on the server for example.
 #[features.${mod_id}]
 #openGLVersion="[3.2,)"
+
+mixinConfigs = "medievaldawnhorizons.mixins.json"
