+++ src/main/resources/medievaldawnhorizons.mixins.json
@@ -1,0 +1,13 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "com.medievaldawngroup.medievaldawnhorizons.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    
+  ],
+  "client": [],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
